Revert to default weight of 1 in vcc2 mode. Add QI_EAGER_THRESHOLD=1000 in vcc3 mode.

+++ vcc/Plugins/Vcc2Plugins/Vcc3/bench/gensx.sh
@@ -3,12 +3,14 @@
 . config.sh
 
 VERSION_OPTS=-3
-SX_DIR=sx-vcc3
 
 if [ "$1" = "-2" ] ; then
   VERSION_OPTS=
-  SX_DIR=sx-vcc2
+  shift
 fi
+SX_DIR="$1"
+
+[ "X$SX_DIR" = X ] && exit 1
 
 rm -rf $SX_DIR
 mkdir $SX_DIR
